✏️ [WV-12] rename : constant 폴더 constants로 변경
